+++ tools/monitor/NodeGenerator.h
@@ -9,7 +9,7 @@
 #pragma once
 #endif /* ACE_LACKS_PRAGMA_ONCE */
 
-#include <QtGui/QGraphicsScene>
+#include <QtWidgets/QGraphicsScene>
 #include <string>
 
 #include "TreeNode.h"